LPD-73239 Reuse AssignProperties in both datasource oveview
